Added a new header for goby::Group and moved it there; removed unneeded mutex from the new middleware::ZeroMQService since this is only called by the InterProcessPortal (which has exactly one per process, so no need for thread safety)
